refactored the astro example to app to abstract the infra out into @tanstack/start
