+++ examples/react/with-astro-ssr/astro.config.mts
@@ -3,6 +3,7 @@ import react from '@astrojs/react'
 import netlify from '@astrojs/netlify/functions'
 import inspect from 'vite-plugin-inspect'
 import bling from '@tanstack/bling/vite'
+import type { AstroIntegration } from 'astro'
 // https://astro.build/config
 export default defineConfig({
   output: 'server',
@@ -17,7 +18,7 @@ function tsr() {
   return {
     name: '@tanstack/astro-plugin-ssr',
     hooks: {
-      'astro:config:setup': async ({ updateConfig }) => {
+      'astro:config:setup': async ({ updateConfig, injectRoute }) => {
         updateConfig({
           vite: {
             plugins: [
@@ -25,18 +26,22 @@ function tsr() {
                 name: '@tanstack/astro-plugin-ssr',
                 enforce: 'post',
                 options(opts) {
-                  return addRollupInput(opts, ['src/app/entry-client.tsx'])
+                  return addRollupInput(opts, ['src/entry-client.tsx'])
                 },
               },
             ],
           },
+        })
+        injectRoute({
+          entryPoint: 'src/entry-server.tsx',
+          pattern: '/[...all]',
         })
       },
       'astro:build:ssr': async ({ manifest }) => {
         console.log(manifest)
       },
     },
-  }
+  } as AstroIntegration
 }
 
 function fromEntries(entries) {
