+++ examples/react/with-astro-ssr/src/.start/routes.ts
@@ -1,0 +1,10 @@
+import { rootRoute } from '../root'
+import { indexRoute } from '../routes/index'
+import { postsRoute } from '../routes/posts'
+import { postsIndexRoute } from '../routes/posts/index'
+import { postIdRoute } from '../routes/posts/$postId'
+
+export const routeTree = rootRoute.addChildren([
+  indexRoute,
+  postsRoute.addChildren([postsIndexRoute, postIdRoute]),
+])
