--- examples/react/with-astro-ssr/src/app/routeTree.ts
@@ -1,10 +1,0 @@
-import { rootRoute } from './routes/__root'
-import { indexRoute } from './routes/index'
-import { postsRoute } from './routes/posts'
-import { postsIndexRoute } from './routes/posts/index'
-import { postIdRoute } from './routes/posts/$postId'
-
-export const routeTree = rootRoute.addChildren([
-  indexRoute,
-  postsRoute.addChildren([postsIndexRoute, postIdRoute]),
-])
