+++ examples/react/with-astro-ssr/src/entry-server.tsx
@@ -1,0 +1,8 @@
+import { createRequestHandler } from '@tanstack/start/server'
+import { createLoaderClient } from './.start/loaders'
+import { routeTree } from './.start/routes'
+
+export const all = createRequestHandler({
+  routeTree,
+  createLoaderClient,
+})
