--- examples/react/with-astro-ssr/src/pages/[...app].ts
@@ -1,4 +1,0 @@
-import type { APIRoute } from 'astro'
-import { createRequestHandler } from '../app/createRequestHandler'
-
-export const all: APIRoute = createRequestHandler()
