--- examples/react/with-astro-ssr/src/pages/astro.astro
@@ -1,4 +1,0 @@
----
----
-
-<div>I am an astro page!</div>
