+++ examples/react/with-astro-ssr/src/routes/posts/$postId.tsx
