+++ examples/react/with-astro-ssr/src/routes/posts/index.tsx
