--- examples/react/with-astro-ssr/src/types.tsx
@@ -1,13 +1,0 @@
-export const FormError = Error
-export const ServerError = Error
-
-export interface FetchEvent {
-  request: Request
-  env: any
-  locals: Record<string, unknown>
-}
-
-export interface ServerFunctionEvent extends FetchEvent {
-  fetch(url: string, init: RequestInit): Promise<Response>
-  // $type: typeof FETCH_EVENT
-}
