+++ packages/bling/server/server.ts
@@ -285,6 +285,7 @@ server$.getHandler = function (route) {
 }
 
 server$.hasHandler = function (route) {
+  console.log(handlers)
   return handlers.has(route)
 }
