+++ packages/start/src/client.ts
@@ -1,0 +1,1 @@
+export * from './StartClient'
