+++ packages/start/src/index.tsx
@@ -1,0 +1,2 @@
+export * from './Scripts'
+export * from './router'
